copper skin first pass
