+++ console/src/App/App.less
@@ -5,8 +5,16 @@
 	// This should prevent iOS from double-tap-to-zoom but... it doesn't completely.
 	touch-action: manipulation;
 
-	.applySkins({
+	.activeAvatar {
+		padding-top: 30px;
 
+		* {
+			height: 150px;
+			width: 150px;
+		}
+	}
+
+	.applySkins({
 		&.copper {
 			background: url(../../assets/themes/copper/background.png) no-repeat left top;
 		}
