+++ console/src/components/WelcomePopup/WelcomePopup.less
@@ -84,24 +84,6 @@
 		}
 	}
 
-	.gridListImageItem {
-		--user-index: 0;
-
-		.image {
-			border-radius: 50%;
-			box-sizing: border-box;
-			position: relative;
-		}
-
-		.caption {
-			font-size: 30px;
-			letter-spacing: 30px;
-			margin-right: -30px;
-			text-transform: uppercase;
-			color: #a47d66;
-		}
-	}
-
 	.smallComponent {
 		margin-top: 12px;
 
@@ -121,6 +103,24 @@
 		font-size: 24px;
 		letter-spacing: 12px;
 		// This refers to theme colors, so we'll get them updated soon after the theme is ready
+		color: #a47d66;
+	}
+}
+
+.gridListImageItem {
+	--user-index: 0;
+
+	.image {
+		border-radius: 50%;
+		box-sizing: border-box;
+		position: relative;
+	}
+
+	.caption {
+		font-size: 30px;
+		letter-spacing: 30px;
+		margin-right: -30px;
+		text-transform: uppercase;
 		color: #a47d66;
 	}
 }
