+++ copilot/src/App/App.less
@@ -16,4 +16,3 @@
 		min-width: 20%;
 	}
 }
-
